Components in subdirectories and include individual styles
