+++ src/components/Form/Checkbox/index.js
@@ -1,0 +1,2 @@
+import Checkbox from './Checkbox'
+export default Checkbox
