+++ src/components/Form/DateControl/index.js
@@ -1,0 +1,2 @@
+import DateControl from './DateControl'
+export default DateControl
