+++ src/components/Form/Email/index.js
@@ -1,0 +1,2 @@
+import Email from './Email'
+export default Email
