+++ src/components/Form/Generic/index.js
@@ -1,0 +1,2 @@
+import Generic from './Generic'
+export default Generic
