+++ src/components/Form/Number/index.js
@@ -1,0 +1,2 @@
+import Number from './Number'
+export default Number
