+++ src/components/Form/Password/index.js
@@ -1,0 +1,2 @@
+import Password from './Password'
+export default Password
