+++ src/components/Form/Radio/index.js
@@ -1,0 +1,2 @@
+import Radio from './Radio'
+export default Radio
