+++ src/components/Form/Telephone/index.js
@@ -1,0 +1,2 @@
+import Telephone from './Telephone'
+export default Telephone
