+++ src/components/Form/Text/index.js
@@ -1,0 +1,2 @@
+import Text from './Text'
+export default Text
