+++ src/components/Form/Textarea/index.js
@@ -1,0 +1,2 @@
+import Textarea from './Textarea'
+export default Textarea
