+++ src/components/Form/Url/index.js
@@ -1,0 +1,2 @@
+import Url from './Url'
+export default Url
